service/requirements.txt
Flask==1.1.1
scikit-learn==0.21.3
sklearn-crfsuite==0.3.6
nltk==3.4.5
